+++ rmk-macro/src/keyboard_config.rs
@@ -2,7 +2,7 @@ use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
 use rmk_config::toml_config::{
     BleConfig, DependencyConfig, KeyboardInfo, KeyboardTomlConfig, LayoutConfig, LightConfig,
-    MatrixConfig, SplitConfig, StorageConfig,
+    MatrixConfig, DirectPinMatrixConfig, SplitConfig, StorageConfig,
 };
 use serde::Deserialize;
 use std::fs;
@@ -83,6 +83,7 @@ pub(crate) struct KeyboardConfig {
 pub(crate) enum BoardConfig {
     Normal(MatrixConfig),
     Split(SplitConfig),
+    DirectPin(DirectPinMatrixConfig),
 }
 
 impl Default for BoardConfig {
@@ -158,7 +159,7 @@ impl KeyboardConfig {
         config.basic = Self::get_basic_info(config.basic, toml_config.keyboard);
 
         // Board config
-        config.board = Self::get_board_config(toml_config.matrix, toml_config.split)?;
+        config.board = Self::get_board_config(toml_config.matrix, toml_config.direct_pin_matrix, toml_config.split)?;
 
         // Layout config
         config.layout = Self::get_layout_from_toml(toml_config.layout)?;
@@ -326,13 +327,15 @@ impl KeyboardConfig {
 
     fn get_board_config(
         matrix: Option<MatrixConfig>,
+        direct_pin_matrix: Option<DirectPinMatrixConfig>,
         split: Option<SplitConfig>,
     ) -> Result<BoardConfig, TokenStream2> {
-        match (matrix, split) {
-            (None, Some(s)) => Ok(BoardConfig::Split(s)),
-            (Some(m), None) => Ok(BoardConfig::Normal(m)),
-            (None, None) => rmk_compile_error!("[matrix] section in keyboard.toml is required for non-split keyboard".to_string()),
-            (Some(_), Some(_)) => rmk_compile_error!("Use either [matrix] or [split] in your keyboard.toml, but NOT BOTH\n-> [matrix] is used to define a normal matrix of non-split keyboard\n-> [split] is used to define a split keyboard".to_string()),
+        match (matrix, direct_pin_matrix, split) {
+            (None, None, Some(s)) => Ok(BoardConfig::Split(s)),
+            (Some(m), None, None) => Ok(BoardConfig::Normal(m)),
+            (None, Some(d), None ) => Ok(BoardConfig::DirectPin(d)),
+            (None, None, None) => rmk_compile_error!("[matrix] section in keyboard.toml is required for non-split keyboard".to_string()),
+            _ => rmk_compile_error!("Use at most one of [matrix], [split], or [direct_pin_martix] in your keyboard.toml!\n-> [matrix] is used to define a normal matrix of non-split keyboard\n-> [split] is used to define a split keyboard\n-> [direct_pin_matrix] is used to define a keyboard of buttons connected directly to pins".to_string()),
         }
     }